Change the publicly_queryable property on the redirect_rule post type to FALSE.
As it stands now, querying a redirect_rule on the front-end will simply display a 404, rather than any sort of "if you weren't previewing this, you'd be redirected to X"-type message.

Setting this value to FALSE also removes the "Preview" button from the Publish meta box, which is determined by `is_post_type_viewable()`.