Fix: Handle empty wildcard requests
Comments in the code suggest the behaviour to allow `/one/` to redirect when Redirect From is `/one/*` and   Redirect To is `/gohere/*` was meant to be, though there was no specific unit test.

This PR adds a unit test, and a fix.

The fix undoes some workaround code related to trailing slashes, and adds specific code to still allow the same behaviour.

Fixes #156.
Props to @mslinnea for an enhanced unit test.